Start Emacs fullscreen
This works in any graphical environment, does not rely on command-line
flags, and does not require modification of Xresources (for X11).

However, Emacs will start at the initial default size before being
maximized. This is OK in my book, because I don't want Emacs maximized
until it has been initialized and is ready for interactive use.