Add sh source block snippet for org-mode
The arguments for this source block are meant to be used in
runbook-style Org buffers in conjunction with my/send-region to evaluate
commands in a separate vterm buffer.

It would be nice to use per-project snippets, as the scenario described
above is not always helpful (e.g. literate config files and
runbook-style Org buffers using ein-bash or jupyter-bash source blocks).