Replace eeowaa-refresh-mode-alist mechanism with revert-buffer-function
Going forward, major modes for non-file-visiting buffers that do not set
a buffer-local `revert-buffer-function' should have their mode hooks
modified to do so (where appropriate).

This might need a bit more work for modes derived from
`tabulated-list-mode', which often has its `revert-buffer-function' set
to `tabulated-list-revert', but where `tabulated-list-revert-hook' is
unset. (So far, I have not encountered such a mode, but it seems
reasonable to expect this.)