Remove error related to transient childframe
It used to be that if you selected the childframe, it would cause the
transient to exit, but the childframe would stay. Now, when you select
the childframe, it will close along with the transient. Although this is
fine, I would prefer to have the ability to click on transient keys in
the childframe to perform actions. (You might want to look into the
`transient-enable-popup-navigation' option.)

I also added some debugging functionality to transient to make it easier
to see what is happening when the transient exits. In the future, I
might want to add more debugging to print out the window's quit-restore
parameter to get a better understanding of how my display-buffer hacks
have been working (and why things break without the hacks).